# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a OpenAPI 2 specification for Mashrooy"
  version: "1.0.0"
  title: "OpenAPI Mashrooy"
host: "localhost:3006"
basePath: "/Mashrooy"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /user:
    get:
      tags:
      - "User"
      summary: "Get All user"
      operationId: "listUsers"
      parameters:
      - name: "rps"
        in: "query"
        description: "Size of the page results in pagination"
        required: false
        type: "integer"
      - name: "rpi"
        in: "query"
        description: "Current page index in pagination"
        required: false
        type: "integer"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "User"
    post:
      tags:
      - "User"
      summary: "Create a list of user"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      parameters:
      - in: "body"
        name: "body"
        description: "Create user List"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/User"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        400:
          description: "Validation Error"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        409:
          description: "Conflict record"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "User"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      parameters:
      - in: "body"
        name: "user"
        description: "Updated user object"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/User"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "User not found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      deprecated: false
      x-swagger-router-controller: "User"
  /user/login:
    post:
      tags:
      - "User"
      summary: "Logs user into the system"
      description: ""
      operationId: "userLogin"
      parameters:
      - in: "body"
        name: "user"
        description: "The user email & Password for login"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
          headers:
            X-access-token:
              type: "string"
              description: "authorization token"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username/password supplied"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "User"
  /user/logout:
    get:
      tags:
      - "User"
      summary: "Logs out current logged in user session"
      description: ""
      operationId: "userLogout"
      parameters:
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "User"
  /user/listdbyFilters:
    post:
      tags:
      - "User"
      summary: "ListUsers By specific Filters"
      operationId: "listUsersByFilters"
      parameters:
      - name: "rps"
        in: "query"
        description: "Size of the page results in pagination"
        required: false
        type: "integer"
      - name: "rpi"
        in: "query"
        description: "Current page index in pagination"
        required: false
        type: "integer"
      - in: "body"
        name: "matchingParams"
        description: "The params to fetch users."
        required: false
        schema:
          $ref: "#/definitions/matchingParams"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        400:
          description: "Validation Error"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "User"
  /user/getUser/{fullName}:
    get:
      tags:
      - "User"
      summary: "Get user by user name"
      operationId: "getUserByName"
      parameters:
      - name: "fullName"
        in: "path"
        description: "The id that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "User not found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "User"
  /user/{id}:
    put:
      tags:
      - "User"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUserById"
      parameters:
      - name: "id"
        in: "path"
        description: "id that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "user"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "User not found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "User"
      summary: "Deletes a User"
      operationId: "deleteUserById"
      parameters:
      - name: "id"
        in: "path"
        description: "User id to delete"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Record Not Found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "User"
...
securityDefinitions:
  loanstore_auth:
    type: "oauth2"
    authorizationUrl: "https://tasaheel.swagger.io/oauth/authorize"
    flow: "implicit"
    scopes:
      write:loans: "modify loans in your account"
      read:loans: "read your loans"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Branch:
    type: "object"
    properties:
      id:
        type: "string"
      branch_code:
        type: "integer"
      branch_name:
        type: "string"
      en_name:
        type: "string"
      districtId:
        type: "string"
      governantId:
        $ref: "#/definitions/Governorate"
      branch_add:
        type: "string"
      en_address:
        type: "string"
      geo:
        $ref: "#/definitions/Branch_geo"
      update_stamp:
        type: "string"
    example:
      branch_code: 0
      geo:
        latitude: "latitude"
        longitude: "longitude"
      districtId: "districtId"
      en_address: "en_address"
      branch_name: "branch_name"
      branch_add: "branch_add"
      en_name: "en_name"
      id: "id"
      governantId: "governantId"
      update_stamp: "update_stamp"
  Governorate:
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "string"
        example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      name:
        type: "object"
        properties: 
           ar: 
             type: string
           en: 
             type: string
  User:
    type: "object"
    properties:
      id:
        type: "string"
      fullName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      branchId:
        $ref: "#/definitions/Branch"
      isActive:
        type: "boolean"
        default: true
      firstLogin:
        type: "boolean"
        default: true
      groups:
        type: "array"
        items:
          $ref: "#/definitions/Group"
      createdBy:
        $ref: "#/definitions/User"
    example:
      branchId:
        branch_code: 0
        geo:
          latitude: "latitude"
          longitude: "longitude"
        districtId: "districtId"
        en_address: "en_address"
        branch_name: "branch_name"
        branch_add: "branch_add"
        en_name: "en_name"
        id: "id"
        governantId: "governantId"
        update_stamp: "update_stamp"
      password: "password"
      phone: "phone"
      firstLogin: true
      fullName: "fullName"
      groups:
      - permissions: "permissions"
        name:
          ar: "ar"
          en: "en"
        id: 0
      - permissions: "permissions"
        name:
          ar: "ar"
          en: "en"
        id: 0
      id: "id"
      isActive: true
      email: "email"
  Group:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        $ref: "#/definitions/Group_name"
      permissions:
        type: "string"
    example:
      permissions: "permissions"
      name:
        ar: "ar"
        en: "en"
      id: 0
...
  Branch_geo:
    type: "object"
    properties:
      longitude:
        type: "string"
      latitude:
        type: "string"
    example:
      latitude: "latitude"
      longitude: "longitude"
...
  matchingParams:
    type: "object"
    properties:
      governantId:
        type: "string"
      districtId:
        type: "string"
      branchId:
        type: "string"
  Group_name:
    type: "object"
    properties:
      ar:
        type: "string"
      en:
        type: "string"
    example:
      ar: "ar"
      en: "en"
...